Introduce Ref::map_val functionality
This change introduces a new associated function to cell::Ref: map_val.
This function returns a RefVal object which in itself contains another
object (as a value, not a reference) that holds a reference to the
borrowed value. This is in contrast to the existing Ref::map function
which returns another Ref. A Ref object, as per its definition, can only
hold a reference to a part of the borrowed data, not a value containing
such a reference.
The most prominent example as to why this functionality may be needed is
probably an iterator. An iterator is a value, but it references the
borrowed data. When having a Ref object and the need to expose an
iterator over the borrowed data, only map_val will allow for such an
exposure.