Reference Rust issue for discussion of problem
The crate provided has duplicated some functionality from the Rust
standard library. Ideally, the Rust authors see the need for the
functionality provided by this crate and agree to accept it upstream.
This possibility is tracked by Rust issue #54776.
With this change we reference this issue in the README to allow
interested developers to easier follow along.