Drop dependency on optin_builtin_traits feature
So far the crate has only been compilable with the nightly toolchain due
to the usage of the optin_builtin_traits unstable feature. Said feature
was thought to be required to ensure that the provided RefCell does not
implement the Sync trait.
However, because our RefCell uses std::cell::UnsafeCell which has the
negative trait bound !Sync associated with it (but in the standard
library where that is allowed unconditionally), there should be no need
to explicitly redefine this negative trait bound.
Hence, with this change we drop this additional definition and, with
that, make the crate compilable with the stable toolchain.