Update README in accordance to Rust Edition 2018 module improvements
With the module improvements made by the Rust Edition 2018 it is no
longer necessary to declare external crates via 'extern crate'. Hence,
this change updates the README to reflect that in the code example. Note
that clients may still use this crate in an Edition 2015 context, in
which the previous syntax is fully supported, but we encourage usage of
Edition 2018.