Add gitlab-ci.yml configuration file
This change adds a configuration file for the Gitlab CI/CD pipeline to
the repository. This file (directly or indirectly) controls the
environment in which to build, what exactly to build, and is also used
to run the test suite. Note that we refrain from running rustfmt on the
code base, as a large part of it is coming from the upstream sources and
we want to minimize the difference to it (but not all of it is properly
formatted according to rustfmt's rules).
With the pipeline in place the change also adds a badge indicating the
status of the build on the project's front page.