Remove implicit Clone constraint from RefVal::map
The entire RefVal impl has a constraint that T implement Clone. When the
map method got added, this constraint implicitly held for it as well.
However, it is unnecessary as there is no requirement to clone anything
as part of this method.
With this change we move the constraint from the impl to the clone
method itself, which, at least right now, is the part of the impl that
requires it.