Update CI/CD pipeline to use Rust 1.35
With an upcoming commit we will bump the baseline of the crate to Rust
1.35. This version contains new functionality that was previously marked
as unstable. Now, because we do not explicitly adjust the doc tests that
come with the original module to use our custom implementation, those
tests will fail to compile with older versions of Rust.
To that end, this change adjusts the CI/CD pipeline to use Rust 1.35 as
the baseline.