Add warning about identified unsafety to README
The crate contains a subtle unsafety for which no solution has been
found so far. As such, users are discouraged from actually using the
crate.
To that end, this change adds a warning to the README. The file is also
updated to reflect the fact that we are no longer aiming to upstream
this functionality.